Move `UiAscl*` types to `ui_loading` crate.
