+++ crate/loading/Cargo.toml
@@ -64,6 +64,7 @@ strum = "0.18.0"
 ui_audio_model = { path = "../ui_audio_model" }
 ui_form_model = { path = "../ui_form_model" }
 ui_label_model = { path = "../ui_label_model" }
+ui_loading = { path = "../ui_loading" }
 ui_menu_item_model = { path = "../ui_menu_item_model" }
 ui_model = { path = "../ui_model" }
 ui_model_spi = { path = "../ui_model_spi" }
