+++ crate/ui_loading/src/lib.rs
@@ -2,6 +2,11 @@
 
 //! Processes UI configuration into the loaded UI model.
 
+pub use self::ui_ascl::{
+    UiAscl, UiAsclCharacterSelection, UiAsclComponents, UiAsclControlSettings, UiAsclForm,
+    UiAsclMapSelection, UiAsclMenu, UiAsclSessionLobby,
+};
 pub use crate::ui_loading_bundle::UiLoadingBundle;
 
+mod ui_ascl;
 mod ui_loading_bundle;
